backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
awscli==1.44.54
awsebcli==3.26
AWSIoTPythonSDK==1.6.0
bcrypt==5.0.0
blessed==1.27.0
boto3==1.42.64
botocore==1.42.64
cement==2.10.14
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
Deprecated==1.3.1
Django==5.2.2
djangorestframework==3.16.1
docutils==0.19
fabric==3.2.2
fastapi==0.133.1
gunicorn==25.1.0
h11==0.16.0
idna==3.11
invoke==2.2.1
jmespath==1.1.0
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
psycopg2-binary
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
semantic-version==2.10.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
starlette==0.52.1
termcolor==2.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.41.0
wcwidth==0.2.14
wrapt==2.1.2
